requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
idna==3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
prometheus-client==0.20.0
